+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - fix a bug: user was not able to reserve at the same time of an event with pre-registration invalidated
 - fix a bug: avoids crash if invoicing_profile has no address associated to it
+- improvement: improves how pay_zen transactions are matched with payment_schedule_items
+- improvement: improves rrule of pay_zen subscriptions
 
 ## v6.3.4 2023 November 23